+++ cmd/psweb/main.go
@@ -34,10 +34,14 @@ import (
 
 const (
 	// App VERSION tag
-	VERSION = "v1.7.6"
-	// Swap Out reserve
-	SWAP_OUT_CHANNEL_RESERVE = 10000
-	// Elements v23.02.03 introduced vsize discount enabled on testnet as default
+	VERSION = "v1.7.7"
+	// Swap Out reserve to deduct from channel local balance
+	SWAP_OUT_CHANNEL_RESERVE = 10_000
+	// https://github.com/ElementsProject/peerswap/pull/304#issuecomment-2303931071
+	SWAP_LBTC_RESERVE = 1_200
+	// Unusable BTC balance
+	ANCHOR_RESERVE = 25_000
+	// assume creatediscountct=1 for mainnet in elements.conf
 	ELEMENTS_DISCOUNTED_VSIZE_VERSION = 230203
 )
 
@@ -1038,12 +1042,12 @@ func convertSwapsToHTMLTable(swaps []*peerswaprpc.PrettyPrintSwap, nodeId string
 		if cost != 0 {
 			totalCost += cost
 			ppm := cost * 1_000_000 / int64(swap.Amount)
-			table += " <span title=\"Swap +profit/-cost, sats. PPM: "
+			table += " <span title=\"Swap "
 
 			if cost < 0 {
-				table += formatSigned(-ppm) + "\">+"
+				table += "profit, sats. PPM: " + formatSigned(-ppm) + "\">+"
 			} else {
-				table += formatSigned(ppm) + "\">"
+				table += "cost, sats. PPM: " + formatSigned(ppm) + "\">"
 			}
 			table += formatSigned(-cost) + "</span>"
 		}
@@ -1129,7 +1133,6 @@ func checkPegin() {
 		// invitation expired
 		ln.ClaimStatus = "No ClaimJoin peg-in is pending"
 		log.Println("Invitation expired from", ln.ClaimJoinHandler)
-		telegramSendMessage("🧬 ClaimJoin Invitation expired")
 
 		ln.ClaimJoinHandler = ""
 		db.Save("ClaimJoin", "ClaimStatus", ln.ClaimStatus)
@@ -1139,13 +1142,11 @@ func checkPegin() {
 	if config.Config.PeginTxId == "" {
 		// send telegram if received new ClaimJoin invitation
 		if peginInvite != ln.ClaimJoinHandler {
-			t := "🧬 There is a ClaimJoin peg-in pending"
-			if ln.ClaimJoinHandler == "" {
-				t = "🧬 ClaimJoin peg-in has ended"
-			} else {
+			t := "⌚ ClaimJoin invitation has expired"
+			if ln.ClaimJoinHandler != "" {
 				duration := time.Duration(10*(ln.JoinBlockHeight-currentBlockHeight)) * time.Minute
-				formattedDuration := time.Time{}.Add(duration).Format("15h 04m")
-				t += ", time limit to join: " + formattedDuration
+				timeLimit := time.Now().Add(duration).Format("3:04 PM")
+				t = "🧬 Invitation to join a confidential peg-in before " + timeLimit
 			}
 			if telegramSendMessage(t) {
 				peginInvite = ln.ClaimJoinHandler
@@ -1319,7 +1320,7 @@ func cacheAliases() {
 // The goal is to spend maximum available liquid
 // To rebalance a channel with high enough historic fee PPM
 func findSwapInCandidate(candidate *SwapParams) error {
-	minAmount := config.Config.AutoSwapThresholdAmount - swapFeeReserveLBTC(10) // assume 10 UTXOs
+	minAmount := config.Config.AutoSwapThresholdAmount - SWAP_LBTC_RESERVE
 	minPPM := config.Config.AutoSwapThresholdPPM
 
 	client, cleanup, err := ps.GetClient(config.Config.RpcHost)
@@ -1500,7 +1501,7 @@ func executeAutoSwap() {
 		return
 	}
 
-	amount = min(amount, satAmount-swapFeeReserveLBTC(10)) // assume 10 UTXOs
+	amount = min(amount, satAmount-SWAP_LBTC_RESERVE)
 
 	// execute swap
 	autoSwapId, err = ps.SwapIn(client, amount, candidate.ChannelId, "lbtc", false)
@@ -1795,14 +1796,14 @@ func advertiseBalances() {
 
 	bitcoinBalance := uint64(ln.ConfirmedWalletBalance(cl))
 	// haircut by anchor reserve
-	if bitcoinBalance >= 25000 {
-		bitcoinBalance -= 25000
+	if bitcoinBalance >= ANCHOR_RESERVE {
+		bitcoinBalance -= ANCHOR_RESERVE
 	}
 
 	liquidBalance := res2.GetSatAmount()
 	// Elements fee bug does not permit sending the whole balance, haircut it
-	if liquidBalance >= 2000 {
-		liquidBalance -= 2000
+	if liquidBalance >= SWAP_LBTC_RESERVE {
+		liquidBalance -= SWAP_LBTC_RESERVE
 	}
 
 	cutOff := time.Now().AddDate(0, 0, -1).Unix() - 120
@@ -1812,7 +1813,7 @@ func advertiseBalances() {
 		maxBalance := uint64(0)
 		if ln.AdvertiseBitcoinBalance || ln.AdvertiseLiquidBalance {
 			for _, ch := range peer.Channels {
-				maxBalance = max(maxBalance, ch.RemoteBalance)
+				maxBalance = max(maxBalance, ch.RemoteBalance-SWAP_OUT_CHANNEL_RESERVE)
 			}
 		}
 
@@ -1931,16 +1932,4 @@ func pollBalances() {
 	if initalPollComplete {
 		log.Println("Polled peers for balances")
 	}
-}
-
-// depends on number of UTXOs
-func swapFeeReserveLBTC(numUTXOs int) uint64 {
-	if hasDiscountedvSize {
-		n := numUTXOs*7 + 20 // better estimate for lots of UTXOs
-		if n < 75 {
-			n = 75 // peerswap assumes 75 sats
-		}
-		return uint64(n)
-	}
-	return 300
 }