+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Versions
 
+## 1.7.7
+
+- Fix Auto Fee not working for channels with no outbound forwards
+- Better indication when all channels with a peer are inactive
+- Catch Bitcoin Core and Elements Core RPC errors
+- Fix telegram bot panic on wrong token or chat id
+- Allow Bitcoin fee rates with 0.01 precision
+- Reserve 1200 sats of LBTC balance to save on Elements fees
+- Show confidential peg-in join time limit on peg-in form
+- Allow updating txid of externally funded peg-in when not found
+
 ## 1.7.6
 
 - Fix BTC to sats rounding bug preventing claim init or join